feat(token-creation): remove create command
BREAKING CHANGE: Remove superfluous create command. ghpat should trigger access token creation.
