+++ src/executables/ghpat.js
@@ -3,8 +3,19 @@
 import program from 'commander';
 
 import pkg from '../../package.json';
+import createPersonalAccessToken from '../createPersonalAccessToken';
+
+const execute = async () => {
+  try {
+    await createPersonalAccessToken();
+  } catch (e) {
+    console.error('😞  Rut ro, an error occurred');
+    console.error(e);
+  }
+};
 
 program.version(pkg.version)
   .description('Create GitHub Personal Access Tokens from the command line')
-  .command('create', 'Create GitHub Personal Access Token')
   .parse(process.argv);
+
+execute();
